refactor: replace global state with Session struct and split netflow into focused files
