+++ replay/replay.go
@@ -10,7 +10,7 @@ import (
 	"context"
 	"encoding/binary"
 	badger "github.com/dgraph-io/badger/v3"
-	"github.com/dmabry/flowgre/flow/netflow"
+	"github.com/dmabry/flowgre/netflow"
 	"github.com/dmabry/flowgre/utils"
 	"log"
 	"net"
