+++ single/single.go
@@ -8,7 +8,7 @@ package single
 import (
 	"encoding/hex"
 	"fmt"
-	"github.com/dmabry/flowgre/flow/netflow"
+	"github.com/dmabry/flowgre/netflow"
 	"github.com/dmabry/flowgre/utils"
 	"log"
 	"net"
